requirements.txt
Django >= 1.3, < 1.4
django-annoying >= 0.7.6
South
selenium
pil